+++ src/network_ops_dashboard/urls.py
@@ -19,6 +19,7 @@ from django.conf import settings
 from django.conf.urls import url, include
 from django.conf.urls.static import static
 from django.contrib.auth import views as auth_views
+from network_ops_dashboard.sdwan.vmanage import views as sdwan_views
 from . import views
 
 urlpatterns = [
@@ -34,8 +35,9 @@ urlpatterns = [
     re_path(r'^dashboard/save-prefs/$', views.dashboard_save_prefs, name='dashboard_save_prefs'),
     re_path(r'^dashboard/toggle/asa_vpn_stats/$', views.dashboard_toggle_asa_vpn_stats, name='dashboard_toggle_asa_vpn_stats'),
     re_path(r'^dashboard/set/asa_vpn_interval/$', views.dashboard_set_asa_vpn_interval, name='dashboard_set_asa_vpn_interval'),
-    re_path(r'dashboard/toggle/email_processing/$', views.dashboard_toggle_email_processing, name='dashboard_toggle_email_processing'),
-    re_path(r'dashboard/set/email_time/$', views.dashboard_set_email_time, name='dashboard_set_email_time'),
+    re_path(r'^dashboard/toggle/email_processing/$', views.dashboard_toggle_email_processing, name='dashboard_toggle_email_processing'),
+    re_path(r'^dashboard/set/email_time/$', views.dashboard_set_email_time, name='dashboard_set_email_time'),
+    re_path(r'^dashboard/sdwan/save/$', views.dashboard_sdwan_settings_save, name='dashboard_sdwan_settings_save'),
     re_path(r'^public_scripts/$', views.public_scripts, name='public_scripts'),
     re_path(r'^protected_media/(?P<path>.*)', views.protected_media, name='protected_media'),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT) + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
@@ -47,6 +49,7 @@ urlpatterns += [
     re_path(r'^asavpn/', include('network_ops_dashboard.asavpn.urls')),
     re_path(r'^apic/', include('network_ops_dashboard.apic.urls')),
     re_path(r'^oncall/', include('network_ops_dashboard.oncall.urls')),
+    re_path(r'^sdwan/vmanage/', include('network_ops_dashboard.sdwan.vmanage.urls')),
     re_path(r'^reports/changes/', include('network_ops_dashboard.reports.changes.urls')),
     re_path(r'^reports/circuits/', include('network_ops_dashboard.reports.circuits.urls')),
     re_path(r'^notices/certexpiry/', include('network_ops_dashboard.notices.certexpiry.urls')),