+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
+## v0.2.3 - 08-21-2025
+- 🆕 Dashboard: Added SD-WAN/vManage Statistics Card.
+- ⚙️ Dashboard: Dynamic card data refreshes automatically.
+- 🖥️ Dashboard/UX: Compacted configuration modal settings for dashboard cards.
+
 ## v0.2.2 - 08-15-2025
 - ⚙️ On-call: Added ability to auto-archive incidents that are "Closed" daily or weekly at a provided time.
 - ✅ On-call: Added status "Archived". Incident log will now show incidents that are "Archived".
 - ✅ Scripts: Lookup NetworkCredential from Inventoy object instead of referring to it due to duplicates and ease.
-- 🗃️ Inventory.object (WIP): Add Forgeignkey fields to NetworkCredential for SSH and REST.
+- 🗃️ Inventory.object: Add Forgeignkey fields to NetworkCredential for SSH and REST.
 - 🖥️ UX: Added superuser views to edit/add SiteSettings & SiteSettingsWebsites.
 - 🖥️ UX: Added staff views to edit/add SiteSecrets to function with network scripts.
 - 🖥️ UX: Added net-admin views to edit/add NetworkCredential & added new model field "username_search_field"